pyproject.toml
[tool.poetry]
name = "sentient-switchblade"
version = "0.2.2"
description = "Unleash Dev Tool Mastery with a Flick of Your Wrist"
authors = ["JensRoland <mail@jensroland.com>"]
license = "MIT"
repository = "https://github.com/jensroland/sentient-switchblade"
homepage = "https://jensroland.com/sentient-switchblade"
readme = "README.md"
packages = [
    { include = "switchbladecli", from = "src" },
]
# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  #! Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
]

[tool.poetry.scripts]
swb = "switchbladecli.cli.__main__:switchbladecli"

[tool.poetry.dependencies]
checksumdir = "^1.2.0"
click = "^8.0.3"
mergedeep = "^1.3.4"
PyGithub = "^1.58.1"
python = "^3.7"
tomlkit = "^0.11.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
pytest-cov = "^4.0.0"
pytest-subprocess = "^1.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
# Directories that are not visited by pytest collector:
norecursedirs =["tests/files", "hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
pythonpath = "src"

[tool.coverage.report]
fail_under = 70
show_missing = true
skip_empty = true
